feat(concurrent): add async future transformers and completion suppliers
This commit adds a new method `blockForValue` to the `Future` class. This method is an alternative to the existing `get()` method and provides unsafe access via blocking to the future's completion value in contexts where the parent context will handle the exception.

The `blockForValue` method allows blocking the current thread and waiting for the future completion to happen. After completion, it returns the completion result object of type T.

If the future completes with an exception, a `FutureExecutionException` is thrown. The actual exception that made the future fail can be obtained using `FutureExecutionException.getCause()`.

The method also supports specifying a timeout for blocking. If the request exceeds the given time interval, a `FutureTimeoutException` is thrown. A timeout of 0 means that the method will block indefinitely.

Additionally, this commit updates all usages of `await()` in favor of using `blockForValue()`.